feat(cli): add --dump option to print SSH URIs
The commit introduces a command-line argument `--dump`, enabling the user to print out all SSH URIs loaded by the script. This feature allows users to get a complete list of the URIs for debugging or inspection purposes. The script will now check if the first argument provided is `--dump` and, if so, it prints the SSH URIs and exits the program, foregoing the UI part of the script.